+++ HISTORY.rst
@@ -1,6 +1,12 @@
 History
 -------
 
+Unreleased
+==================
+
+* Re-implement encoding for base32 and base64, as the implementation was buggy
+* Add extensive tests for all encodings
+
 0.1.0 (2017-09-02)
 ==================